requirements.txt
requests-walmart-auth~=1.0.1
